Refactor TypeStruct
Split out each field into a TypeFld, which is also interned, and a collection of TypeFlds, which are also interned.  Enforce access via getters.  Use an Enum instead of a const int for field mods.
Reduce the number of factory methods.

Hoping to lower the complexity of TypeStruct, prior to allowing args to match in any order for structural subtyping.

Code compilers, never executed.