One more pass to get nilable right
All HM tests pass.
Add a nilable type-var; its always a nilable-version of a plain type-var.  If the type-var ever becomes a nilable type (any ground term, struct or function), then it collapses into being a nilable version of that term.
'0' turns into a nilable-leaf: (? T)
Unify of Q:(? X) and R:(? Y) is Q>>R and X>>Y.
Unify of Q:(? X) and Y always unifies Q and Y (keeping the nilable), but makes a shallow copy of Y without the nilable which unifies with X.
Struct unification has to deal with the underlying 'this' unifying into a nilable variant of self.